+++ app/javascript/react_app/components/bird/details_modal.jsx
