+++ app/javascript/react_app/components/bird/modal.jsx
