+++ app/javascript/react_app/components/group/group_list.jsx
@@ -2,14 +2,14 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import { capitalize } from '../utils';
-import { fetchGroups, sortGroups, setGroupListScrollPos } from '../actions';
+import { capitalize } from '../../utils';
+import { fetchGroups, sortGroups, setGroupListScrollPos } from '../../actions';
 
-import Group from '../components/group';
+import Group from './group';
 import SearchBar from './search_bar';
-import GroupHeader from '../components/group_header';
-import PageTitle from '../components/page_title';
-import OptionsBar from '../components/options_bar';
+import GroupHeader from '../shared/group_header';
+import PageTitle from '../shared/page_title';
+import OptionsBar from '../shared/options_bar';
 
 class GroupList extends Component {
   componentDidMount() {
