+++ app/javascript/react_app/components/group/search_bar.jsx
@@ -3,8 +3,8 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { Link } from 'react-router-dom';
 
-import { fetchSearchBirds, clearSearchBirds } from '../actions';
-import { hashify } from '../utils';
+import { fetchSearchBirds, clearSearchBirds } from '../../actions';
+import { hashify } from '../../utils';
 
 class SearchBar extends Component {
   constructor(props) {
