+++ app/javascript/react_app/components/settings/settings.jsx
@@ -2,10 +2,10 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import { saveSettings, setFlashMessage } from '../actions';
+import { saveSettings, setFlashMessage } from '../../actions';
 
-import PageTitle from '../components/page_title';
-import BackLink from '../components/back_link';
+import PageTitle from '../shared/page_title';
+import BackLink from '../shared/back_link';
 
 class Settings extends Component {
   constructor(props) {
