+++ app/javascript/react_app/components/shared/flash_message.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import { clearFlashMessage } from '../actions';
+import { clearFlashMessage } from '../../actions';
 
 class FlashMessage extends Component {
   componentDidMount() {
