+++ app/javascript/react_app/components/shared/group_header.jsx
