+++ app/javascript/react_app/components/shared/navbar.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import logo from '../../../assets/images/transparent_64.png';
+import logo from '../../../../assets/images/transparent_64.png';
 
 const Navbar = () => (
   <div className="navbar navbar-expand-sm navbar-light">
